+++ src/Recyclarr.Cli/Console/Commands/BaseCommandSettings.cs
@@ -12,4 +12,9 @@ public class BaseCommandSettings : CommandSettings
     [Description("Show debug logs in console output.")]
     [UsedImplicitly(ImplicitUseKindFlags.Assign)]
     public bool Debug { get; init; }
+
+    [CommandOption("--app-data")]
+    [Description("Custom path to the application data directory")]
+    [UsedImplicitly(ImplicitUseKindFlags.Assign)]
+    public string? AppData { get; init; }
 }